requirements.txt
numpy>=1.13.3
torch>=1.0.0
scipy>=1.0.0
pandas>=0.21.0
visdom>=0.1.05
click>=6.7
xlrd>=1.0.0
cytoolz>=0.9.0.1
tqdm>=4.11.2
scikit-learn>=0.19.1
flake8>=3.6.0
tensorboardX>=1.2
setuptools>=40.2.0
torchvision>=0.2.1
seaborn>=0.9.0
pytest>=3.8.0
pytest-cov>=2.6.0
codecov>=2.0.15
Faker==1.0.5
pyyaml==6.0.1
rapidfuzz==3.4.0
pyarrow==12.0.1
